Separate classes/functions into smaller files. Implement basic elevation filtering and compatibility with IUCN range maps, albeit not polished or well tested yet
